refactor: create SortableItem file
